move function in main to cmd/root.go
